load-balancer/requirements.txt
fastapi==0.115.0
httpx==0.27.2
uvicorn[standard]==0.32.0
gunicorn==23.0.0
python-multipart==0.0.12
python-dotenv==1.0.0